feat: feature parity with arena implementation
