+++ cmd/run.c
@@ -101,14 +101,6 @@ int run(const run_opts_t OPTIONS) {
   tryCLI(arenaCreate(OPTIONS.ast_memory), ast_arena,
          "unable to allocate interpreter memory");
 
-  arena_t *scratch_arena = nullptr;
-  tryCLI(arenaCreate(OPTIONS.temp_memory / 2), scratch_arena,
-         "unable to allocate transient memory");
-
-  arena_t *result_arena = nullptr;
-  tryCLI(arenaCreate(OPTIONS.temp_memory / 4), result_arena,
-         "unable to allocate transient memory");
-
   vm_options_t vm_options = {
       .environment_size = OPTIONS.environment_size,
       .vm_size = OPTIONS.temp_memory / 4,
@@ -119,8 +111,6 @@ int run(const run_opts_t OPTIONS) {
   do {
     memset(statement_buffer, 0, (size_t)file_length);
     arenaReset(ast_arena);
-    arenaReset(scratch_arena);
-    arenaReset(result_arena);
     readStatement(file_length, statement_buffer, file_buffer, &file_offset);
 
     if (strlen(statement_buffer) == 0)
@@ -135,7 +125,9 @@ int run(const run_opts_t OPTIONS) {
     tryRun(parse(ast_arena, tokens, &line_offset, &depth), syntax_tree);
 
     if (syntax_tree) {
-      tryRun(evaluate(syntax_tree, machine->global));
+      value_t *result;
+      tryRun(evaluate(syntax_tree, machine->global), result);
+      valueDestroy(&result);
     }
   } while (file_offset < file_length);
 
@@ -144,8 +136,6 @@ int run(const run_opts_t OPTIONS) {
   deallocSafe(&statement_buffer);
   deallocSafe(&file_buffer);
 
-  arenaDestroy(&result_arena);
-  arenaDestroy(&scratch_arena);
   arenaDestroy(&ast_arena);
   return 0;
 }
