+++ lifp/evaluate.h
@@ -4,20 +4,9 @@
 #include "value.h"
 
 // Evaluate a node in the context of a given environment.
-//
-// The result arena is where all the values required for the result will be
-// stored. This is where values that need to survive across `evaluate` calls
-// need to be stored.
-//
-// The scratch arena is used for intermediate allocations and it's expected to
-// be wiped at the end of each call. No values to be persisted should be stored
-// there.
+// Returns pointer to a value that the caller needs to free.
 result_value_ref_t evaluate(node_t *, environment_t *);
 
 // Invokes a closure with the specified arguments and environment.
-//
-// This function executes the given closure, passing in the provided arguments,
-// using the specified scratch arena for temporary allocations, and within the
-// given environment. The result of the closure execution is stored in the
-// location pointed to by `result`.
+// Returns pointer to a value that the caller needs to free.
 result_value_ref_t invokeClosure(value_t *, value_array_t *);
